requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyascii==0.3.2
asttokens==2.4.1
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
bangla==0.0.2
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
cryptography==43.0.1
confection==0.1.5
contourpy==1.3.0
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dateparser==1.1.8
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
g2pkk==0.1.2
greenlet==3.0.3
grpcio==1.66.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
hangul-romanize==0.1.0
huggingface-hub==0.24.6
idna==3.8
inflect==7.3.1
ipython==8.27.0
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.7
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
nltk==3.9.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
packaging==24.1
pandas==1.5.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pynndescent==0.5.13
pyparsing==3.1.4
pypinyin==0.52.0
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
srt==3.5.3
stack-data==0.6.3
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
tqdm==4.66.5
trainer==0.0.36
traitlets==5.14.3
transformers==4.44.2
TTS==0.22.0
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.2
vosk==0.3.45
waitress
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==13.0.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.11
